End of /NAMES list was missing in one place. fixed compile warning in ltrace
git-svn-id: http://svn.ratbox.org/svnroot/trunk@1695 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/m_ltrace.c
@@ -129,7 +129,7 @@ int     m_ltrace(struct Client *cptr,
       name = get_client_name(acptr, FALSE);
       ip = inetntoa((char*) &acptr->ip);
 
-      c_class = get_client_class(acptr);
+      c_class = (int)get_client_class(acptr);
 
       if (IsAnyOper(acptr))
         { 
@@ -185,7 +185,7 @@ int     m_ltrace(struct Client *cptr,
       name = get_client_name(acptr, FALSE);
       ip = inetntoa((const char*) &acptr->ip);
 
-      c_class = get_client_class(acptr);
+      c_class = (int)get_client_class(acptr);
       
       switch(acptr->status)
         {
